refactor(data_operations): split numeric_source into per-CFW modules
numeric_source.py bundled all five compute frameworks (duckdb, sqlite,
pandas, polars, pyarrow) into one module, which forced lazy per-helper
imports and broke the repo's one-module-per-framework convention (cf.
pandas_helpers.py, polars_mode_helpers.py).

Split into duckdb/sqlite/pandas/polars/pyarrow _numeric_source.py, each
importing only its own framework at the top level. Descriptor bodies and
DUCKDB_NUMERIC_PREFIXES are moved verbatim; backends and the #214
single-source guards repoint to the per-framework modules. No behavior
change.
