+++ mloda/community/feature_groups/data_operations/row_preserving/point_arithmetic/duckdb_point_arithmetic.py
@@ -10,7 +10,7 @@ from mloda_plugins.compute_framework.base_implementations.duckdb.duckdb_relation
 from mloda_plugins.compute_framework.base_implementations.sql.sql_utils import quote_ident
 
 from mloda.community.feature_groups.data_operations.errors import unsupported_op_error
-from mloda.community.feature_groups.data_operations.numeric_source import duckdb_non_numeric_descriptor
+from mloda.community.feature_groups.data_operations.duckdb_numeric_source import duckdb_non_numeric_descriptor
 from mloda.community.feature_groups.data_operations.row_preserving.point_arithmetic.base import (
     PointArithmeticFeatureGroup,
 )
