+++ mloda/community/feature_groups/data_operations/row_preserving/point_arithmetic/pandas_point_arithmetic.py
@@ -8,7 +8,7 @@ from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.pandas.dataframe import PandasDataFrame
 
 from mloda.community.feature_groups.data_operations.errors import unsupported_op_error
-from mloda.community.feature_groups.data_operations.numeric_source import pandas_non_numeric_descriptor
+from mloda.community.feature_groups.data_operations.pandas_numeric_source import pandas_non_numeric_descriptor
 from mloda.community.feature_groups.data_operations.row_preserving.point_arithmetic.base import (
     ARITHMETIC_OPERATIONS,
     PointArithmeticFeatureGroup,
