+++ mloda/community/feature_groups/data_operations/row_preserving/scalar_arithmetic/pyarrow_scalar_arithmetic.py
@@ -9,7 +9,7 @@ from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.pyarrow.table import PyArrowTable
 
 from mloda.community.feature_groups.data_operations.errors import unsupported_op_error
-from mloda.community.feature_groups.data_operations.numeric_source import pyarrow_non_numeric_descriptor
+from mloda.community.feature_groups.data_operations.pyarrow_numeric_source import pyarrow_non_numeric_descriptor
 from mloda.community.feature_groups.data_operations.row_preserving.scalar_arithmetic.base import (
     ARITHMETIC_OPERATIONS,
     ScalarArithmeticFeatureGroup,
